+++ docs/versionhistory.rst
@@ -19,6 +19,8 @@ APScheduler, see the :doc:`migration section <migration>`.
 * Fixed ``BlockingScheduler`` and ``BackgroundScheduler`` shutdown hanging after the user has
   erroneously tried to start it twice
 * Fixed memory leak when coroutine jobs raise exceptions (due to reference cycles in tracebacks)
+* Fixed inability to schedule wrapped functions with extra arguments when the wrapped function
+  cannot accept them but the wrapper can (original PR by Egor Malykh)
 
 
 3.6.3